fix: add endpoint for listing tenants
BREAKING CHANGES: Tenant service is renamed
